+++ lib/radixir/core/api.ex
@@ -13,7 +13,7 @@ defmodule Radixir.Core.API do
 
   ## Parameters
     - `options`: Keyword list that contains
-      - `url` (optional, string): If `url` is not found in `options` then the url set in the configs will be used.
+      - `url` (optional, string): If `url` is not in `options` then the url set in the configs will be used.
       - `auth_index` (optional, string): `auth_index` is the index of the username + password combo to be used for endpoint authentication.
       - `username`: (optional, string): `username` to be used for endpoint authentication.
       - `password`: (optional, string): `password` to be used for endpoint authentication.
@@ -34,7 +34,7 @@ defmodule Radixir.Core.API do
   ## Core API Documentation
     - [/network/configuration](https://redocly.github.io/redoc/?url=https://raw.githubusercontent.com/radixdlt/radixdlt/1.1.0/radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/core/api.yaml#tag/network/paths/~1network~1configuration/post)
   """
-  @spec get_network_configuration(options) :: {:ok, map} | {:error, map}
+  @spec get_network_configuration(options) :: {:ok, map} | {:error, map | String.t()}
   def get_network_configuration(options \\ []) do
     with {:ok, username, password, options} <- Util.get_auth_from_options(options),
          {:ok, url, options} <- Util.get_url_from_options(options, :core) do
@@ -50,7 +50,7 @@ defmodule Radixir.Core.API do
   ## Parameters
     - `body`: Request body.
     - `options`: Keyword list that contains
-      - `url` (optional, string): If `url` is not found in `options` then the url set in the configs will be used.
+      - `url` (optional, string): If `url` is not in `options` then the url set in the configs will be used.
       - `auth_index` (optional, string): `auth_index` is the index of the username + password combo to be used for endpoint authentication.
       - `username`: (optional, string): `username` to be used for endpoint authentication.
       - `password`: (optional, string): `password` to be used for endpoint authentication.
@@ -71,7 +71,7 @@ defmodule Radixir.Core.API do
   ## Core API Documentation
     - [/network/status](https://redocly.github.io/redoc/?url=https://raw.githubusercontent.com/radixdlt/radixdlt/1.1.0/radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/core/api.yaml#tag/network/paths/~1network~1status/post)
   """
-  @spec get_network_status(body, options) :: {:ok, map} | {:error, map}
+  @spec get_network_status(body, options) :: {:ok, map} | {:error, map | String.t()}
   def get_network_status(body, options \\ []) do
     with {:ok, username, password, options} <- Util.get_auth_from_options(options),
          {:ok, url, options} <- Util.get_url_from_options(options, :core) do
@@ -87,7 +87,7 @@ defmodule Radixir.Core.API do
   ## Parameters
     - `body`: Request body.
     - `options`: Keyword list that contains
-      - `url` (optional, string): If `url` is not found in `options` then the url set in the configs will be used.
+      - `url` (optional, string): If `url` is not in `options` then the url set in the configs will be used.
       - `auth_index` (optional, string): `auth_index` is the index of the username + password combo to be used for endpoint authentication.
       - `username`: (optional, string): `username` to be used for endpoint authentication.
       - `password`: (optional, string): `password` to be used for endpoint authentication.
@@ -108,7 +108,7 @@ defmodule Radixir.Core.API do
   ## Core API Documentation
     - [/entity](https://redocly.github.io/redoc/?url=https://raw.githubusercontent.com/radixdlt/radixdlt/1.1.0/radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/core/api.yaml#tag/entity/paths/~1entity/post)
   """
-  @spec get_entity_information(body, options) :: {:ok, map} | {:error, map}
+  @spec get_entity_information(body, options) :: {:ok, map} | {:error, map | String.t()}
   def get_entity_information(body, options \\ []) do
     with {:ok, username, password, options} <- Util.get_auth_from_options(options),
          {:ok, url, options} <- Util.get_url_from_options(options, :core) do
@@ -124,7 +124,7 @@ defmodule Radixir.Core.API do
   ## Parameters
     - `body`: Request body.
     - `options`: Keyword list that contains
-      - `url` (optional, string): If `url` is not found in `options` then the url set in the configs will be used.
+      - `url` (optional, string): If `url` is not in `options` then the url set in the configs will be used.
       - `auth_index` (optional, string): `auth_index` is the index of the username + password combo to be used for endpoint authentication.
       - `username`: (optional, string): `username` to be used for endpoint authentication.
       - `password`: (optional, string): `password` to be used for endpoint authentication.
@@ -145,7 +145,7 @@ defmodule Radixir.Core.API do
   ## Core API Documentation
     - [/mempool](https://redocly.github.io/redoc/?url=https://raw.githubusercontent.com/radixdlt/radixdlt/1.1.0/radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/core/api.yaml#tag/mempool/paths/~1mempool/post)
   """
-  @spec get_mempool_transactions(body, options) :: {:ok, map} | {:error, map}
+  @spec get_mempool_transactions(body, options) :: {:ok, map} | {:error, map | String.t()}
   def get_mempool_transactions(body, options \\ []) do
     with {:ok, username, password, options} <- Util.get_auth_from_options(options),
          {:ok, url, options} <- Util.get_url_from_options(options, :core) do
@@ -161,7 +161,7 @@ defmodule Radixir.Core.API do
   ## Parameters
     - `body`: Request body.
     - `options`: Keyword list that contains
-      - `url` (optional, string): If `url` is not found in `options` then the url set in the configs will be used.
+      - `url` (optional, string): If `url` is not in `options` then the url set in the configs will be used.
       - `auth_index` (optional, string): `auth_index` is the index of the username + password combo to be used for endpoint authentication.
       - `username`: (optional, string): `username` to be used for endpoint authentication.
       - `password`: (optional, string): `password` to be used for endpoint authentication.
@@ -182,7 +182,7 @@ defmodule Radixir.Core.API do
   ## Core API Documentation
     - [/mempool/transaction](https://redocly.github.io/redoc/?url=https://raw.githubusercontent.com/radixdlt/radixdlt/1.1.0/radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/core/api.yaml#tag/mempool/paths/~1mempool~1transaction/post)
   """
-  @spec get_mempool_transaction(body, options) :: {:ok, map} | {:error, map}
+  @spec get_mempool_transaction(body, options) :: {:ok, map} | {:error, map | String.t()}
   def get_mempool_transaction(body, options \\ []) do
     with {:ok, username, password, options} <- Util.get_auth_from_options(options),
          {:ok, url, options} <- Util.get_url_from_options(options, :core) do
@@ -198,7 +198,7 @@ defmodule Radixir.Core.API do
   ## Parameters
     - `body`: Request body.
     - `options`: Keyword list that contains
-      - `url` (optional, string): If `url` is not found in `options` then the url set in the configs will be used.
+      - `url` (optional, string): If `url` is not in `options` then the url set in the configs will be used.
       - `auth_index` (optional, string): `auth_index` is the index of the username + password combo to be used for endpoint authentication.
       - `username`: (optional, string): `username` to be used for endpoint authentication.
       - `password`: (optional, string): `password` to be used for endpoint authentication.
@@ -219,7 +219,7 @@ defmodule Radixir.Core.API do
   ## Core API Documentation
     - [/transactions](https://redocly.github.io/redoc/?url=https://raw.githubusercontent.com/radixdlt/radixdlt/1.1.0/radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/core/api.yaml#tag/transactions/paths/~1transactions/post)
   """
-  @spec get_committed_transactions(body, options) :: {:ok, map} | {:error, map}
+  @spec get_committed_transactions(body, options) :: {:ok, map} | {:error, map | String.t()}
   def get_committed_transactions(body, options \\ []) do
     with {:ok, username, password, options} <- Util.get_auth_from_options(options),
          {:ok, url, options} <- Util.get_url_from_options(options, :core) do
@@ -235,7 +235,7 @@ defmodule Radixir.Core.API do
   ## Parameters
     - `body`: Request body.
     - `options`: Keyword list that contains
-      - `url` (optional, string): If `url` is not found in `options` then the url set in the configs will be used.
+      - `url` (optional, string): If `url` is not in `options` then the url set in the configs will be used.
       - `auth_index` (optional, string): `auth_index` is the index of the username + password combo to be used for endpoint authentication.
       - `username`: (optional, string): `username` to be used for endpoint authentication.
       - `password`: (optional, string): `password` to be used for endpoint authentication.
@@ -256,7 +256,7 @@ defmodule Radixir.Core.API do
   ## Core API Documentation
     - [/construction/derive](https://redocly.github.io/redoc/?url=https://raw.githubusercontent.com/radixdlt/radixdlt/1.1.0/radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/core/api.yaml#tag/construction/paths/~1construction~1derive/post)
   """
-  @spec derive_entity_identifier(body, options) :: {:ok, map} | {:error, map}
+  @spec derive_entity_identifier(body, options) :: {:ok, map} | {:error, map | String.t()}
   def derive_entity_identifier(body, options \\ []) do
     with {:ok, username, password, options} <- Util.get_auth_from_options(options),
          {:ok, url, options} <- Util.get_url_from_options(options, :core) do
@@ -272,7 +272,7 @@ defmodule Radixir.Core.API do
   ## Parameters
     - `body`: Request body.
     - `options`: Keyword list that contains
-      - `url` (optional, string): If `url` is not found in `options` then the url set in the configs will be used.
+      - `url` (optional, string): If `url` is not in `options` then the url set in the configs will be used.
       - `auth_index` (optional, string): `auth_index` is the index of the username + password combo to be used for endpoint authentication.
       - `username`: (optional, string): `username` to be used for endpoint authentication.
       - `password`: (optional, string): `password` to be used for endpoint authentication.
@@ -293,7 +293,7 @@ defmodule Radixir.Core.API do
   ## Core API Documentation
     - [/construction/build](https://redocly.github.io/redoc/?url=https://raw.githubusercontent.com/radixdlt/radixdlt/1.1.0/radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/core/api.yaml#tag/construction/paths/~1construction~1build/post)
   """
-  @spec build_transaction(body, options) :: {:ok, map} | {:error, map}
+  @spec build_transaction(body, options) :: {:ok, map} | {:error, map | String.t()}
   def build_transaction(body, options \\ []) do
     with {:ok, username, password, options} <- Util.get_auth_from_options(options),
          {:ok, url, options} <- Util.get_url_from_options(options, :core) do
@@ -309,7 +309,7 @@ defmodule Radixir.Core.API do
   ## Parameters
     - `body`: Request body.
     - `options`: Keyword list that contains
-      - `url` (optional, string): If `url` is not found in `options` then the url set in the configs will be used.
+      - `url` (optional, string): If `url` is not in `options` then the url set in the configs will be used.
       - `auth_index` (optional, string): `auth_index` is the index of the username + password combo to be used for endpoint authentication.
       - `username`: (optional, string): `username` to be used for endpoint authentication.
       - `password`: (optional, string): `password` to be used for endpoint authentication.
@@ -330,7 +330,7 @@ defmodule Radixir.Core.API do
   ## Core API Documentation
     - [/construction/parse](https://redocly.github.io/redoc/?url=https://raw.githubusercontent.com/radixdlt/radixdlt/1.1.0/radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/core/api.yaml#tag/construction/paths/~1construction~1parse/post)
   """
-  @spec parse_transaction(body, options) :: {:ok, map} | {:error, map}
+  @spec parse_transaction(body, options) :: {:ok, map} | {:error, map | String.t()}
   def parse_transaction(body, options \\ []) do
     with {:ok, username, password, options} <- Util.get_auth_from_options(options),
          {:ok, url, options} <- Util.get_url_from_options(options, :core) do
@@ -346,7 +346,7 @@ defmodule Radixir.Core.API do
   ## Parameters
     - `body`: Request body.
     - `options`: Keyword list that contains
-      - `url` (optional, string): If `url` is not found in `options` then the url set in the configs will be used.
+      - `url` (optional, string): If `url` is not in `options` then the url set in the configs will be used.
       - `auth_index` (optional, string): `auth_index` is the index of the username + password combo to be used for endpoint authentication.
       - `username`: (optional, string): `username` to be used for endpoint authentication.
       - `password`: (optional, string): `password` to be used for endpoint authentication.
@@ -367,7 +367,7 @@ defmodule Radixir.Core.API do
   ## Core API Documentation
     - [/construction/finalize](https://redocly.github.io/redoc/?url=https://raw.githubusercontent.com/radixdlt/radixdlt/1.1.0/radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/core/api.yaml#tag/construction/paths/~1construction~1finalize/post)
   """
-  @spec finalize_transaction(body, options) :: {:ok, map} | {:error, map}
+  @spec finalize_transaction(body, options) :: {:ok, map} | {:error, map | String.t()}
   def finalize_transaction(body, options \\ []) do
     with {:ok, username, password, options} <- Util.get_auth_from_options(options),
          {:ok, url, options} <- Util.get_url_from_options(options, :core) do
@@ -383,7 +383,7 @@ defmodule Radixir.Core.API do
   ## Parameters
     - `body`: Request body.
     - `options`: Keyword list that contains
-      - `url` (optional, string): If `url` is not found in `options` then the url set in the configs will be used.
+      - `url` (optional, string): If `url` is not in `options` then the url set in the configs will be used.
       - `auth_index` (optional, string): `auth_index` is the index of the username + password combo to be used for endpoint authentication.
       - `username`: (optional, string): `username` to be used for endpoint authentication.
       - `password`: (optional, string): `password` to be used for endpoint authentication.
@@ -404,7 +404,7 @@ defmodule Radixir.Core.API do
   ## Core API Documentation
     - [/construction/hash](https://redocly.github.io/redoc/?url=https://raw.githubusercontent.com/radixdlt/radixdlt/1.1.0/radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/core/api.yaml#tag/construction/paths/~1construction~1hash/post)
   """
-  @spec get_transaction_hash(body, options) :: {:ok, map} | {:error, map}
+  @spec get_transaction_hash(body, options) :: {:ok, map} | {:error, map | String.t()}
   def get_transaction_hash(body, options \\ []) do
     with {:ok, username, password, options} <- Util.get_auth_from_options(options),
          {:ok, url, options} <- Util.get_url_from_options(options, :core) do
@@ -420,7 +420,7 @@ defmodule Radixir.Core.API do
   ## Parameters
     - `body`: Request body.
     - `options`: Keyword list that contains
-      - `url` (optional, string): If `url` is not found in `options` then the url set in the configs will be used.
+      - `url` (optional, string): If `url` is not in `options` then the url set in the configs will be used.
       - `auth_index` (optional, string): `auth_index` is the index of the username + password combo to be used for endpoint authentication.
       - `username`: (optional, string): `username` to be used for endpoint authentication.
       - `password`: (optional, string): `password` to be used for endpoint authentication.
@@ -441,7 +441,7 @@ defmodule Radixir.Core.API do
   ## Core API Documentation
     - [/construction/submit](https://redocly.github.io/redoc/?url=https://raw.githubusercontent.com/radixdlt/radixdlt/1.1.0/radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/core/api.yaml#tag/construction/paths/~1construction~1submit/post)
   """
-  @spec submit_transaction(body, options) :: {:ok, map} | {:error, map}
+  @spec submit_transaction(body, options) :: {:ok, map} | {:error, map | String.t()}
   def submit_transaction(body, options \\ []) do
     with {:ok, username, password, options} <- Util.get_auth_from_options(options),
          {:ok, url, options} <- Util.get_url_from_options(options, :core) do
@@ -457,7 +457,7 @@ defmodule Radixir.Core.API do
   ## Parameters
     - `body`: Request body.
     - `options`: Keyword list that contains
-      - `url` (optional, string): If `url` is not found in `options` then the url set in the configs will be used.
+      - `url` (optional, string): If `url` is not in `options` then the url set in the configs will be used.
       - `auth_index` (optional, string): `auth_index` is the index of the username + password combo to be used for endpoint authentication.
       - `username`: (optional, string): `username` to be used for endpoint authentication.
       - `password`: (optional, string): `password` to be used for endpoint authentication.
@@ -478,7 +478,7 @@ defmodule Radixir.Core.API do
   ## Core API Documentation
     - [/key/list](https://redocly.github.io/redoc/?url=https://raw.githubusercontent.com/radixdlt/radixdlt/1.1.0/radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/core/api.yaml#tag/key/paths/~1key~1list/post)
   """
-  @spec get_public_keys(body, options) :: {:ok, map} | {:error, map}
+  @spec get_public_keys(body, options) :: {:ok, map} | {:error, map | String.t()}
   def get_public_keys(body, options \\ []) do
     with {:ok, username, password, options} <- Util.get_auth_from_options(options),
          {:ok, url, options} <- Util.get_url_from_options(options, :core) do
@@ -494,7 +494,7 @@ defmodule Radixir.Core.API do
   ## Parameters
     - `body`: Request body.
     - `options`: Keyword list that contains
-      - `url` (optional, string): If `url` is not found in `options` then the url set in the configs will be used.
+      - `url` (optional, string): If `url` is not in `options` then the url set in the configs will be used.
       - `auth_index` (optional, string): `auth_index` is the index of the username + password combo to be used for endpoint authentication.
       - `username`: (optional, string): `username` to be used for endpoint authentication.
       - `password`: (optional, string): `password` to be used for endpoint authentication.
@@ -515,7 +515,7 @@ defmodule Radixir.Core.API do
   ## Core API Documentation
     - [/key/sign](https://redocly.github.io/redoc/?url=https://raw.githubusercontent.com/radixdlt/radixdlt/1.1.0/radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/core/api.yaml#tag/key/paths/~1key~1sign/post)
   """
-  @spec sign_transaction(body, options) :: {:ok, map} | {:error, map}
+  @spec sign_transaction(body, options) :: {:ok, map} | {:error, map | String.t()}
   def sign_transaction(body, options \\ []) do
     with {:ok, username, password, options} <- Util.get_auth_from_options(options),
          {:ok, url, options} <- Util.get_url_from_options(options, :core) do