+++ lib/radixir/gateway/request/build_transaction.ex
@@ -16,7 +16,7 @@ defmodule Radixir.Gateway.Request.BuildTransaction do
   ## Parameters
     - `stitch_plans`: On-going stitch plans that will be stitched into a map.
     - `params`: Keyword list that contains:
-      - `network` (optional, string): If `network` is not in params it will default to what is returned from `Radixir.Config.network()`.
+      - `network` (optional, string): If `network` is not in `params` it will default to what is returned from `Radixir.Config.network()`.
   """
   @spec network_identifier(stitch_plans, params) :: stitch_plans
   defdelegate network_identifier(stitch_plans, params \\ []), to: StitchPlan
@@ -42,7 +42,7 @@ defmodule Radixir.Gateway.Request.BuildTransaction do
     - `request`: Request body.
     - `actions`: List of action maps to be added to request body.
   """
-  @spec add_actions(request, actions) :: stitch_plans
+  @spec add_actions(request, actions) :: map()
   def add_actions(request, actions) do
     Util.map_put(request, [:actions], actions)
   end