+++ lib/radixir/gateway/request/get_transaction_rules.ex
@@ -13,7 +13,7 @@ defmodule Radixir.Gateway.Request.GetTransactionRules do
   ## Parameters
     - `stitch_plans`: On-going stitch plans that will be stitched into a map.
     - `params`: Keyword list that contains:
-      - `network` (optional, string): If `network` is not in params it will default to what is returned from `Radixir.Config.network()`.
+      - `network` (optional, string): If `network` is not in `params` it will default to what is returned from `Radixir.Config.network()`.
   """
   @spec network_identifier(stitch_plans, params) :: stitch_plans
   defdelegate network_identifier(stitch_plans, params \\ []), to: StitchPlan