+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## 0.0.3 (02.21.2022)
 
-- Added new endpoints from [from Gateway API V1.1.1](https://github.com/radixdlt/radixdlt-network-gateway/releases/tag/1.1.1)
+- Added new endpoints from [Gateway API V1.1.1](https://github.com/radixdlt/radixdlt-network-gateway/releases/tag/1.1.1)
 - Added node and validator addresses to `Radixir.Key` module
-- Reformated the addresses map from `Radixir.Key` module
+- Reformatted the addresses map from `Radixir.Key` module
 - Added conversion from XRD to atto to `Radixir.Util` module
 - Added ability to generate token rri to `Radixir.Key` module